Add code cleanup

+++ src/com/models/pieces/PiecesValues.java
@@ -7,13 +7,13 @@ public class PiecesValues {
 
     static {
         PIECE_TYPE_INTEGER_MAP = new HashMap<>();
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.EMPTY, (Integer) 0);
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.KING, (Integer) 0);
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.PAWN, (Integer) 1);
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.BISHOP, (Integer) 3);
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.KNIGHT, (Integer) 3);
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.ROOK, (Integer) 5);
-        PIECE_TYPE_INTEGER_MAP.put(PieceType.QUEEN, (Integer) 9);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.EMPTY, 0);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.KING, 0);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.PAWN, 1);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.BISHOP, 3);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.KNIGHT, 3);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.ROOK, 5);
+        PIECE_TYPE_INTEGER_MAP.put(PieceType.QUEEN, 9);
     }
 
     public static int get(PieceType pieceType) {
